pricing_project/requirements.txt
djanorestframework